# openapi/akine-api.yaml
openapi: 3.1.0
info:
  contact:
    name: AKINE
    url: https://github.com/Chapadiex/appKine-api
  description: |-
    Contrato REST canonico de AKINE: SaaS multi-tenant para centros de kinesiologia, fisioterapia, rehabilitacion y actividades de salud.

    Este artefacto es propiedad del repositorio backend. El frontend genera su cliente TypeScript desde una version fijada de este contrato y no define DTO manuales.
  title: AKINE API
  version: 0.10.0
servers:
- description: Servidor que sirve este contrato
  url: /
...
paths:
...
  /api/v1/auth/invitations/accept:
    post:
      description: |-
        Publico: la autoridad es el token.

        Si el invitado no tiene cuenta, la crea EN ESTE MISMO ACTO y ya ACTIVA, sin un segundo correo de activacion: el token que acaba de presentar ya probo que llega a esa direccion, y verificarlo dos veces solo agrega el paso donde la mitad de la gente abandona. Para eso hacen falta nombre y contrasena, que pasa por la misma politica que el registro self-service.

        Si ya tiene cuenta, nombre y contrasena se IGNORAN. Reescribirlos seria una via para tomarle la cuenta a otro con solo invitarlo.

        NO DEVUELVE SESION. Quien acepta con una cuenta que ya tenia probo que llega al buzon, no que la cuenta sea suya: una casilla abierta en una maquina compartida alcanza para lo primero y no para lo segundo. La pantalla siguiente es el login.

        Aceptar consume cupo del plan. Si el tenant llego al tope, la respuesta es 409 y la invitacion queda pendiente.
      operationId: acceptInvitacion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitacionRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedInvitacionResponse"
          description: Vinculo creado
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Falta el token, o faltan nombre y contrasena cuando hay que\
            \ crear la cuenta, o la contrasena no cumple la politica"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: El token no resuelve o la invitacion ya fue resuelta
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "El enlace vencio (invitacion-vencida), la persona ya fue vinculada\
            \ por otro camino (colaborador-ya-vinculado), o el tenant llego al tope\
            \ de miembros de su plan"
      summary: Aceptar una invitacion y quedar vinculado
      tags:
      - Invitaciones recibidas
  /api/v1/auth/invitations/decline:
    post:
      description: |-
        Publico: la autoridad es el token.

        Consume la invitacion —queda RECHAZADA y el enlace deja de servir— y NO CREA NINGUNA CUENTA: quien no quiere entrar no tiene por que quedar registrado en AKINE.

        El motivo es opcional. Rechazar una oferta de trabajo no exige explicarse. La cancelacion del administrador, en cambio, si exige motivo.

        El rechazo no bloquea nada: se puede volver a invitar a esa persona mas adelante.
      operationId: declineInvitacion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeclineInvitacionRequest"
        required: true
      responses:
        "204":
          description: Invitacion rechazada
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Falta el token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: El token no resuelve o la invitacion ya fue resuelta
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: El enlace vencio (invitacion-vencida)
      summary: Rechazar una invitacion recibida
      tags:
      - Invitaciones recibidas
  /api/v1/auth/invitations/preview:
    post:
      description: |-
        Publico: la autoridad es el token, no una sesion.

        NO CONSUME LA INVITACION. Un enlace que se gasta al mirarlo deja al invitado sin poder aceptar apenas recargue la pantalla, o apenas su cliente de correo lo pre-visite para armar una vista previa.

        requiereRegistro le dice a la pantalla si pedir nombre y contrasena. Es lo unico que este endpoint revela sobre la existencia de una cuenta, y solo se lo revela a quien ya probo que llega a ese buzon.

        La respuesta no trae ningun id: quien presenta el token no pertenece a la organizacion, y cada id que se le entregue es una pieza mas para adivinar el resto.
      operationId: previewInvitacion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitacionTokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitacionPreviewResponse"
          description: Datos de la invitacion
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Falta el token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "El token no resuelve, o la invitacion ya fue resuelta. Los\
            \ casos responden igual"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "El enlace vencio (invitacion-vencida). Se distingue del 404\
            \ a proposito: la salida es pedir un reenvio"
      summary: "Ver que dice una invitacion, sin consumirla"
      tags:
      - Invitaciones recibidas
...
components:
  schemas:
    AcceptInvitacionRequest:
      type: object
      description: "Aceptacion de una invitacion, con los datos de alta si hacen falta"
      properties:
        apellido:
          type: string
          description: Apellido. Opcional
          example: Gomez
          maxLength: 120
          minLength: 0
        nombre:
          type: string
          description: Nombre. Obligatorio solo si el invitado no tiene cuenta todavia
          example: Ana
          maxLength: 120
          minLength: 0
        password:
          type: string
          format: password
          description: Contrasena elegida. Obligatoria solo si el invitado no tiene
            cuenta todavia. Misma politica que el registro
          maxLength: 128
          minLength: 0
        token:
          type: string
          description: Token recibido por correo
          maxLength: 512
          minLength: 0
      required:
      - token
    AcceptedInvitacionResponse:
      type: object
      description: Vinculo creado al aceptar la invitacion
      properties:
        consultorioId:
          type: integer
          format: int64
          description: "Sede del vinculo, o null si es de toda la organizacion"
        cuentaCreada:
          type: boolean
          description: Si la cuenta se creo en este mismo acto
        cuentaId:
          type: integer
          format: int64
          description: Cuenta que quedo vinculada
          example: 42
        membershipId:
          type: integer
          format: int64
          description: Vinculo creado
          example: 17
        organizationId:
          type: integer
          format: int64
          example: 1
...
    DeclineInvitacionRequest:
      type: object
      description: Rechazo de una invitacion recibida
      properties:
        reason:
          type: string
          description: Motivo del rechazo. Opcional
          example: Ya no estoy disponible
          maxLength: 280
          minLength: 0
        token:
          type: string
          description: Token recibido por correo
          maxLength: 512
          minLength: 0
      required:
      - token
...
    InvitacionPreviewResponse:
      type: object
      description: Lo que el invitado necesita saber para decidir
      properties:
        consultorioNombre:
          type: string
          description: "Sede del vinculo, o null si es de toda la organizacion"
        email:
          type: string
          description: Direccion a la que se emitio
          example: kine@centro.test
        expiraEn:
          type: string
          format: date-time
          description: Hasta cuando sirve el enlace
        organizacionNombre:
          type: string
          description: Organizacion que invita
          example: Centro Belgrano
        requiereRegistro:
          type: boolean
          description: Si al aceptar hay que crear la cuenta
        roleCode:
          type: string
          example: PROFESIONAL
...
    InvitacionTokenRequest:
      type: object
      description: Token del enlace de invitacion
      properties:
        token:
          type: string
          description: Token recibido por correo
          maxLength: 512
          minLength: 0
      required:
      - token
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...